+++ Assets/Plugins/Substance Extensions/Examples/Scripts/Runtime/Runtime/RuntimeRenderExample.cs
@@ -8,14 +8,16 @@ namespace SOS.SubstanceExtensions.Examples
     /// </summary>
     public class RuntimeRenderExample : MonoBehaviour
     {
-        [SerializeField]
+        [SerializeField, RuntimeGraphOnly, Tooltip("Substance to render.")]
         private SubstanceGraphSO substance = null;
-        [SerializeField]
+        [SerializeField, Tooltip("If true, rendering will be done asynchronously on a separate thread. This typically prevents frame drops from rendering a substance.")]
+        private bool renderAsync = false;
+        [SerializeField, Tooltip("Input values to set on the target substance.")]
         private SubstanceParameterValue[] parameters = new SubstanceParameterValue[0];
         [Header("Outputs")]
-        [SerializeField]
+        [SerializeField, Tooltip("Parent transform for spawned output previews.")]
         private RectTransform outputParent = null;
-        [SerializeField]
+        [SerializeField, Tooltip("Prefab for generated output textures.")]
         private SubstanceOutputPreview previewPrefab = null;
         [SerializeField]
         private SubstanceOutput[] outputs = new SubstanceOutput[0];
@@ -34,19 +36,50 @@ namespace SOS.SubstanceExtensions.Examples
             //Dispose the native graph to free up resources.
             substance.EndRuntimeEditing(nativeGraph);
 
+            GenerateOutputs();
+        }
+
+
+        private async void RenderSubstanceAsync()
+        {
+            //Get a SubstanceNativeGraph for the render operations and initialize the substance graph for runtime.
+            SubstanceNativeGraph nativeGraph = substance.BeginRuntimeEditing();
+
+            //Set input values on the native graph.
+            await nativeGraph.SetInputValuesAsync(parameters);
+
+            //Render the native graph asynchronously and update output textures on the substance graph.
+            await substance.RenderAsync(nativeGraph);
+
+            //Dispose the native graph to free up resources.
+            substance.EndRuntimeEditing(nativeGraph);
+
+            GenerateOutputs();
+        }
+
+
+        private void GenerateOutputs()
+        {
             //Generate preview outputs.
-            for(int i=0; i < outputs.Length; i++)
+            for (int i = 0; i < outputs.Length; i++)
             {
                 SubstanceOutputPreview outputPreview = Instantiate(previewPrefab, outputParent);
 
-                outputPreview.Initialize(outputs[i].Name, substance.GetOutputMap(outputs[i].Name));
+                outputPreview.Initialize(outputs[i].Name, substance.GetOutputTexture(outputs[i].Name));
             }
         }
 
 
         private void Start()
         {
-            RenderSubstance();
+            if (renderAsync)
+            {
+                RenderSubstanceAsync();
+            }
+            else
+            {
+                RenderSubstance();
+            }
         }
     }
 }